+++ mycli/main.py
@@ -995,6 +995,9 @@ class MyCli(object):
 @click.option('--ssh-port', default=22, help='Port to connect to ssh server.')
 @click.option('--ssh-password', help='Password to connect to ssh server.')
 @click.option('--ssh-key-filename', help='Private key filename (identify file) for the ssh connection.')
+@click.option('--ssh-config-path', help='Path to ssh configuration.',
+              default=os.path.expanduser('~') + '/.ssh/config')
+@click.option('--ssh-config-host', help='Host to connect to ssh server reading from ssh configuration.')
 @click.option('--ssl-ca', help='CA file in PEM format.',
               type=click.Path(exists=True))
 @click.option('--ssl-capath', help='CA directory.')
@@ -1016,6 +1019,8 @@ class MyCli(object):
               help='Use DSN configured into the [alias_dsn] section of myclirc file.')
 @click.option('--list-dsn', 'list_dsn', is_flag=True,
         help='list of DSN configured into the [alias_dsn] section of myclirc file.')
+@click.option('--list-ssh-config', 'list_ssh_config', is_flag=True,
+              help='list ssh configurations in the ssh config (requires paramiko).')
 @click.option('-R', '--prompt', 'prompt',
               help='Prompt format (Default: "{0}").'.format(
                   MyCli.default_prompt))
@@ -1048,7 +1053,7 @@ def cli(database, user, host, port, socket, password, dbname,
         ssl_ca, ssl_capath, ssl_cert, ssl_key, ssl_cipher,
         ssl_verify_server_cert, table, csv, warn, execute, myclirc, dsn,
         list_dsn, ssh_user, ssh_host, ssh_port, ssh_password,
-        ssh_key_filename):
+        ssh_key_filename, list_ssh_config, ssh_config_path, ssh_config_host):
     """A MySQL terminal client with auto-completion and syntax highlighting.
 
     \b
@@ -1084,6 +1089,31 @@ def cli(database, user, host, port, socket, password, dbname,
                 click.secho("{} : {}".format(alias, value))
             else:
                 click.secho(alias)
+        sys.exit(0)
+    if list_ssh_config:
+        if not paramiko:
+            click.secho(
+                "This features requires paramiko. Please install paramiko and try again.",
+                err=True, fg='red'
+            )
+            exit(1)
+        try:
+            ssh_config = paramiko.config.SSHConfig().from_path(ssh_config_path)
+        except paramiko.ssh_exception.ConfigParseError as err:
+            click.secho('Invalid SSH configuration file. '
+                        'Please check the SSH configuration file.',
+                        err=True, fg='red')
+            exit(1)
+        except FileNotFoundError as e:
+            click.secho(str(e), err=True, fg='red')
+            exit(1)
+        for host in ssh_config.get_hostnames():
+            if verbose:
+                host_config = ssh_config.lookup(host)
+                click.secho("{} : {}".format(
+                    host, host_config.get('hostname')))
+            else:
+                click.secho(host)
         sys.exit(0)
     # Choose which ever one has a valid value.
     database = dbname or database
@@ -1134,6 +1164,32 @@ def cli(database, user, host, port, socket, password, dbname,
             host = uri.hostname
         if not port:
             port = uri.port
+
+    if ssh_config_host:
+        if not paramiko:
+            click.secho(
+                "This features requires paramiko. Please install paramiko and try again.",
+                err=True, fg='red'
+            )
+            exit(1)
+        try:
+            ssh_config = paramiko.config.SSHConfig().from_path(ssh_config_path)
+        except paramiko.ssh_exception.ConfigParseError as err:
+            click.secho('Invalid SSH configuration file. '
+                        'Please check the SSH configuration file.',
+                        err=True, fg='red')
+            exit(1)
+        except FileNotFoundError as e:
+            click.secho(str(e), err=True, fg='red')
+            exit(1)
+        ssh_config = ssh_config.lookup(ssh_config_host)
+        ssh_host = ssh_host if ssh_host else ssh_config.get('hostname')
+        ssh_user = ssh_user if ssh_user else ssh_config.get('user')
+        if ssh_config.get('port') and ssh_port == 22:
+            # port has a default value, overwrite it if it's in the config
+            ssh_port = int(ssh_config.get('port'))
+        ssh_key_filename = ssh_key_filename if ssh_key_filename else ssh_config.get(
+            'identityfile', [''])[0]
 
     if not paramiko and ssh_host:
         click.secho(